backend/requirements.txt
asgiref==3.8.1
Django==4.2.16
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
djangorestframework==3.12.4
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0